+++ requirements.txt
@@ -4,3 +4,5 @@ google-genai
 pandas
 langchain_google_genai
 langchain_core
+pytesseract
+pdf2image
